+++ tests/test-crate/lib.rs
@@ -1,4 +1,7 @@
-#![allow(rustdoc::redundant_explicit_links, rustdoc::broken_intra_doc_links)]
+#![allow(rustdoc::redundant_explicit_links)]
+#![allow(rustdoc::broken_intra_doc_links)]
+#![feature(trait_alias)]
+#![feature(extern_types)]
 //! - A shortcut link: [`Vec`]!
 //! - An inline link: [`String`](std::string::String)!
 //! - A reference: [`ThinRope`].
@@ -16,6 +19,32 @@
 //! - A link to a struct that is re-exported: [`Reexport`].
 //! - A link to a struct that is re-exported with `#[doc(inline)]`: [`ReexportInline`].
 //! - A link to a struct that is re-exported from a private module: [`ReexportPrivate`].
+//!
+//! - A link to a module: [`my_module`]
+//! - A link to an extern crate: [`alloc`]
+//! - A link to a use: [`MyStructUse`]
+//! - A link to a union: [`MyUnion`]
+//! - A link to a struct: [`MyStruct`]
+//! - A link to a struct field: [`MyStruct::my_field`]
+//! - A link to an enum: [`MyEnum`]
+//! - A link to a variant: [`MyEnum::MyVariant`]
+//! - A link to a function: [`my_function`]
+//! - A link to a trait: [`MyTrait`]
+//! - A link to a trait alias: [`IntoString`]
+//! - A link to an impl block is not possible
+//! - A link to a type alias: [`MyStructAlias`]
+//! - A link to a constant: [`MY_CONSTANT`]
+//! - A link to a static: [`MY_STATIC`]
+//! - A link to an extern type: [`MyExternType`]
+//! - A link to a macro: [`my_macro`]
+//! - A link to a proc macro: [`phf_macros::phf_map`]
+//! - A link to a primitive: [`i32`]
+//! - A link to an associated constant: [`MyTrait::MY_ASSOCIATED_CONSTANT`]
+//! - A link to an associated type: [`MyTrait::MyAssociatedType`]
+//! - A link to a proc macro attribute is not possible?
+//! - A link to a proc macro derive: [`Debug`]
+//! - A link to a keyword is not possible
+//! - A link to a method: [`MyStruct::my_method`]
 //!
 //! [`ThinRope`]: String
 //!
@@ -91,4 +120,38 @@ pub use reexport_private::ReexportPrivate;
 
 mod reexport_private {
     pub struct ReexportPrivate;
+}
+
+// here come tests to check that we can link to any item kind
+
+pub mod my_module {}
+pub extern crate alloc;
+pub use MyStruct as MyStructUse;
+pub union MyUnion {
+    _x: u8,
+}
+pub struct MyStruct {
+    pub my_field: i32,
+}
+impl MyStruct {
+    pub fn my_method(&self) {}
+}
+pub enum MyEnum {
+    MyVariant,
+}
+#[macro_export]
+macro_rules! my_macro {
+    () => {};
+}
+pub fn my_function() {}
+pub trait MyTrait {
+    const MY_ASSOCIATED_CONSTANT: i32 = 0;
+    type MyAssociatedType;
+}
+pub trait MyTraitAlias = Into<String>;
+pub type MyStructAlias = MyStruct;
+pub const MY_CONSTANT: i32 = 0;
+pub static MY_STATIC: i32 = 0;
+unsafe extern "C" {
+    pub type MyExternType;
 }