+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- next-header -->
 ## [Unreleased]
 
+### Added
+
+- Add support for links of all item kinds (enum variant, macro attribute, etc.)
+
 ## [0.2.1] - 2025-07-10
 
 ### Added